Apply Johannes Erdfelt's big-concurrency.patch
To: qmail@list.cr.yp.to
Subject: Re: mail volume
Date: 4 Aug 1999 20:41:00 -0700

On Thu, Aug 05, 1999, richard@illuin.demon.co.uk <richard@illuin.demon.co.uk> wrote:
> On Wed, 4 Aug 1999, Daemeon Reiydelle wrote:
>
> > (2.6 or later). There may be limitations within e.g. qmail-[lr]spawn
> > about how many children it can manage. I am not working with that code
> > right know so I don't know. Anyone?
>
> This is what people have been trying to say -- the protocol between
> qmail-Xspawn and qmail-send only passes a single byte for the delivery
> attempt back in the status messages. if you want to increase the maximum
> number above 256 one has to modify qmail-send and the common code in
> qmail-Xspawn. making it a short should allow up to 2**16 concurrency
> remotes.
>
> **CAUTION** if you do this one should realise that qmail-send might try to
> open 64K connections to the /same/ host because it doesn't maintain a
> per-domain concurrency. this is distinctly Unfriendly. I produced some
> code for qmail to do this, but when I asked my ISP if i could open >>1024
> connections to one of their mail relays for testing they were less than
> enthusiastic... (the code is on my desktop system somewhere between here
> and Austin where I'm moving to next week, so I can't email it, and without
> testing it I won't email it. the changes to up the concurrency are fairly
> straightforward, the once for a per-domain concurrency are non-trivial)

This is the patch that I use at suse.com. We do almost 1 million
messages a day with this patch and concurrencyremote set to 400.

This patch comes with the standard disclaimer. No warranty, it may not
work, etc. But it works for me :)

It's also not pretty. It's against qmail-1.03+verh-0.02 (the ezmlm patch
l and h patch). So the offsets may be off a little bit.

JE

From: "Fred Lindberg" <lindberg@id.wustl.edu>
To: "nelson@crynwr.com" <nelson@crynwr.com>
Cc: "qmail@list.cr.yp.to" <qmail@list.cr.yp.to>
Subject: Solved: qmail-bigrem limits linux
Date: Thu, 12 Aug 1999 13:41:18 -0500

Dear Russell,

Thanks for your input. The limiting factor for my redhat linux 2.2.5
installation was a per user process limit of 256. The reason I
sometimes got 256 or 257 concurrency is that reporting is not exactly
synchronous with forking.

Thus, in addition to increasing the number of file handles, you need to
rebuild the kernel after editing:

/usr/src/linux/include/linux/tasks.h NR_TASKS from 512 to e.g. 2048.
Per-user limit defined to half this on the line below.

There also appears to be a bug that limits the number of per process
file handles to 1024. There are "unofficial" patches for this, but
AFAIK, not integrated into the official kernel.

If anyone has more insight into this, I'd love to hear it. With the
patch, our P100/64MB does very well at a concurrencyremote of 400.
Since performance was limited by concurrency (many slow/dead clients)
we got considerably better throughput.

-Sincerely, Fred

(Frederik Lindberg, Infectious Diseases, WashU, St. Louis, MO, USA)