Let qmail-smtpd reject admin-defined recipients.
Original: https://www.soffian.org/downloads/qmail/qmail-smtpd.patch

This patch modifies qmail-smtpd to check for an environment variable,
$RCPTCHECK. If it is set, for each 'rcpt to:' that qmail-smtpd receives,
it will fork/exec $RCPTCHECK. $RCPTCHECK is run in the same environment
as qmail-smtpd (Note that qmail-smtpd changes its CWD to qmail-home
[typically /var/qmail] at startup). Additionally $SENDER is set to the
envelope from (mail from:) and $RECIPIENT is set to the envelope
recipient (for the current rcpt to:). Based on the return code (exit
value) of $RCPTCHECK, the rcpt to: address will either be accepted or
rejected as follows:

- 100: recipient is rejected with "553 sorry, no mailbox here by that
  name. (#5.1.1)"
- 111: connection is dropped with a temporary error "421 unable to
  verify recipient (#4.3.0)"
- 120: connection is dropped with a temporary error "421 unable to
  execute recipient check (#4.3.0)"
- All others: recipient is accepted.

120 is used internally (by the patch) if $RCPTCHECK cannot be executed.

Useful for validating recipient to addresses with an arbitrary program
(such as fastforward in '-n' mode).