Reject many {Q,S}MTP recipients that would bounce.
Original: qmail-1.03-realrcptto-2006.12.10.patch

The qmail-realrcptto patch copies logic from qmail-send, qmail-lspawn,
qmail-getpw, and qmail-local into qmail-smtpd and qmail-qmtpd, so that
if a local delivery (i.e., one for a domain in /var/qmail/control/locals
or virtualdomains) would eventually bounce due to a missing .qmail file,
then that recipient address is rejected during the SMTP or QMTP protocol
conversation. There are other qmail patches around that do similar jobs
(badrcptto, etc.); the focus of this patch is to get this functionality
with no additional administrative effort, rather than qmail-smtpd's
running speed. You just set up your .qmail files, as you would have to
do anyway, and the rest is automatic, though slower than a CDB lookup.

Addresses which use the default delivery instructions are never rejected
by this patch, because they would never be bounced due to the lack of a
.qmail file.

This patch is less effective when there are .qmail-default files, or
when qmaild does not have sufficient filesystem permissions to stat()
users' .qmail files, since in those cases, qmail-smtpd cannot know that
the delivery would later bounce due to the lack of a .qmail file. For
example, I'm told that this patch is not useful for virtual domains
managed with vpopmail, since an applicable .qmail-default file always
exists. (I'm not familiar with vpopmail myself.) The patch may still be
useful for other local or virtual domains on the same server, if they
are not managed by vpopmail.