detect qmail system uid/gid at runtime.
rename and extend auto_usera.h to auto_users.h and include all system
accounts, not just the alias user.

Instead of looking up the uid/gid of each account in conf-users at
build time, record their names in the same manner as the alias user.

remove auto-gid.c and auto-uid.c, replacing them with runtime
equivalent gid.c and uid.c.  Look up the uid and gid values in or
about the main function of the five qmail programs that use user
or group ids, and the installer.

Link substdio.a in to qmail-start, as runtime error messages are
produced when an account cannot be found.

This allows qmail to be built on a host that does not have user and
group system accounts.