split install in to instpackage, instchown.
Introduce a new phony target, package, that runs instpackage.
Unlike the setup target, which runs install, this target will
ignore the qmail system accounts and install all files and
directories as the current (i.e., non-root) user.

It is intended to be run on a build host, which may not have
the qmail system accounts, for the purpose of packaging a
release.

instchown sets the user, group, and mode of the files and
directories in a qmail install.  Along with instpackage, it
splits the install program in to the part run on the build host
(instpackage) and the part run on the install host (instchown).