get rid of str_len() and just use the default strlen() instead
Basically 2 things could happen here:

-the custom implementation is used, which is probably slower than the one
 provided by the compiler
-the compiler or linker notices that this is just strlen() and replaces it with
 the normal one

In both cases we are better of just removing it altogether.

It would in theory be possible that this implementation is faster than what your
compiler or libc offers. But than your system is already slow, this few calls
for your mailer daemon will not matter then anyway.