Apply Russell Nelson's big-todo.103.patch.
---Note to readers: After you apply this patch, your queue will no
longer be readable in its current state.  So, either install this into
/var/qmail2 and run both qmails until the queue is emptied on the old
one, or else accept that your current queue will have to be destroyed
and ``rm -rf /var/qmail/queue; make setup''.---

From: Bruce Guenter <bruceg@em.ca>
To: Russell Nelson <nelson@crynwr.com>
Subject: Missing piece of your big-todo patch
Date: Tue, 3 Aug 1999 08:12:11 -0600

Greetings.

While testing the performance of a big todo directory, I observed that
qmail-qstat miscounted the number of messages in the todo directory.
I've appended the necessary patch to yours.
--
Bruce Guenter <bruceg@em.ca>                       http://em.ca/~bruceg/

Another issue that can come into play when email comes into
your server extremely quickly is that if several thousand email
messages accumulate in the todo queue, an inefficient filesystem
that has trouble with large directories may restrict the speed
of todo processing significantly. The best way to address this
is to use a better filesystem, but that's not always possible,
depending on your situation. Of course, qmail already worked
around this kind of problem in the rest of the queue by using
a hash system to reduce the number of message files per directory
in the queue. For whatever reason, this workaround wasn't applied
to the todo queue. Russ Nelson wrote a patch to make qmail use
that hashing system in the todo portion of the queue as well.
This patch is referred to as the "big-todo" patch. There's no
reason to apply this patch unless you really really need it,
because getting a better filesystem (or enabling the right
features on the one you're already using) is the best option;
this patch can require that your queue be rebuilt, and so
applying it to a running system can be a bit of a pain.