Apply Christopher K. Davis's oversize DNS packet patch.
Originally named qmail-103.patch

Sometimes (rarely) the answer to a DNS query is larger than 512
bytes (the max that qmail allows (which was based on the UDP DNS
protocol definition (RFC 1025, section 4.2.1))). This is not a
widespread occurrence, yet, but can and does happen from time to
time. This patch allows DNS packets to be as big as the maximum
DNS response size, but does not waste memory if you never see
one that's that big.