remove shar target from Makefile.
This target restates the contents of FILES as dependencies on the
shar target, a shell archive.  Shell archives are no longer used
for distributing software.  While the command could be updated to
use an alternative distribution format, like tape archive (tar),
that the rule must restate the contents of FILES means any
modification to FILES must also be restated here, and in a way
that is more likely to cause a merge conflict.