remove phony targets from TARGETS, add them as .PHONY prerequisites.
For version of make that support it (GNU make, OpenBSD, NetBSD, and
others), the .PHONY special target will arrange to always run a
rule whose name appears as a prerequisite, regardless of whether the
target exists or is up to date.

With these names no longer in TARGETS, they will not be removed if
they are created, though the build will no longer consider them.

The build will not create these targets as files under any normal
circumstance, so a make that does not recognize the .PHONY special
target should silently ignore it, but a user would have to manually
remove any files with the name of a phony target they otherwise
created.