CI: build & package before creating user accounts.
Move the build to before we create qmail system accounts, verifying
the compile works without them.  Do a trial package in to a non-
root owned directory along with a package build as root in to
/var/qmail.

After creating qmail system accounts, run instchown over the prior
package build, then remove the install and redo it using the regular
install program.

This tests that the build and package step can be done without qmail
system accounts.