make the queue directory world readable
This affects only the queue directory (i.e. /var/qmail/queue) itself, but not
any of the subdirectories. The idea is that one can open that directory and then
is able to call e.g. fstatvfs() on it. This is e.g. used by Qsmtp to check if a
given mail size (by SIZE extension) is too large for the queue and avoid
needless network traffic.