Make "ar" and "ranlib" configurable like "cc" and "ld" are
DJB generate compilation tools by concatenating

- warning header
- a file with CC='' LD=''
- make-<file>.sh with the echo 'things to add in <file>'

The end compilation tool <file> gets gen generated by executing make-<file>
which produces a short script with the commands to execute directly.

The make-makelib script did not have the same treatment as the others,
as AR and RANLIB was not configurable.  This patch gives the ability to
configure them the same way.

This also helps with avoiding a special case in #107.