generate compile, load, makelib directly with cat
Currently, the make-<file>.sh scripts generate a make-<file> that is
executed to obtain the <file> used for compiling.

The end ./compile script goes from this:

	#!/bin/sh
	# WARNING: This file was auto-generated. Do not edit!
	exec cc -O2 -c "$@"

to this:

	#!/bin/sh
	# WARNING: This file was auto-generated. Do not edit!
	CC='cc -O2'
	LD='cc -s'
	AR='ar'
	RANLIB='ranlib'
	exec $CC -c "$@"

So this is a tradeoff between having shorter end-result ./compile,
./load and ./makelib scripts, and having a simpler build process:
all .sh are now generated only with cat(1), and there is no script
that generate a script anymore.

An exception is ./auto-ccld.sh, which still takes its content from conf-*
config files.