remove alloc.c, wrapper over malloc and use malloc directly
As discussed in #112, there are little advantages remaining today
to keep a custom allocator implementation in qmail.

The file alloc.c and alloc_re.c are removed entirely, to only keep
macros in alloc.h.

The Makefile is edited to not compile either of these file, and
alloc.a is not built anymore, so removed from everywhere as well.

alloc() is not a function-style macro, because alloc is used as
a function pointer by some cdb code:

	$ grep ,alloc *.c
	cdbmake_add.c:int cdbmake_add(cdbm,h,p,alloc)
	cdbmake_add.c:int cdbmake_split(cdbm,alloc)
	cdbmss.c:  if (!cdbmake_add(&c->cdbm,h,c->pos,alloc)) return -1;
	cdbmss.c:  if (!cdbmake_split(&c->cdbm,alloc)) return -1;