genalloc: unfold tail code
This was doing 2 assignments, a function call and a return on the same line.
This had the additional slight bug that if the allocation failed it returned 0,
but the allocation length in the struct passed in was already changed to the new
value. This would not do any harm on subsequent calls as the pointer was still
NULL and only that is checked.