make now() an inline function
This is a thin wrapper about time(), so just let it do that as inline function.
It can't be a macro, as it is called as a function but without arguments. Now
it also returns time_t, which avoids losing precision on platforms where time_t
is larger than long.