qmail-remote: add infrastructure for EHLO parsing
Implementing extensions for qmail-remote often requires the usage of EHLO and
checking if the server supports the extension. Every patch therefore usually
comes with it's own version of EHLO parsing. Add yet another one that the
patches can easily hook into. For the moment it will just send mails using
ESMTP instead of SMTP for basically all cases, but not use any of the new
information.