qmail-remote: verify the SMTP result code of remote servers much stricter
The server reply must be of the form

  (\d\d\d-[^\r\n]*\r\n)*
  \d\d\d [^\r\n]*\r\n*

Until now it was neither verified that the first 3 characters were digits, nor
that the following character was really SPACE or MINUS. In case any of these
restrictions violated return a high, invalid, SMTP code, that will cause all
callers to return permanent error.

While at it make sure noone can send an absurdly low number as status code,
and return an error in that case, too.