add noreturn attribute to a couple of functions
When building with -Wextra there are a lot of false positives for fallthroughs,
which happen because the compiler does not detect that these functions in fact
never return. Add an attribute if building with gcc. According to the
documentation it is supported since version 2.5, which is way older than
anything one can reasonably expect this to work on.