Let QMAILQUEUE wrapper return custom error strings.
Original: qmail-queue-custom-error-v2.netqmail-1.05.patch

A common QMAILQUEUE use case is to prevent some messages from entering
the queue. We extend the qmail-queue(8) interface so that when rejecting
a message, a custom string can optionally be communicated to the caller
(for instance, qmail-smtpd). When this is desired, write to fd 6:

    "Dthis is a custom fatal error message"
      or
    "Zthis is a custom temporary failure message"

and exit 82.

An earlier version of this patch has been around since 2005. In 2007, it
switched from fd 4 to 6 in order to play well with qmail-qfilter.