Handle DNS packets up to max EDNS response size.
Original: qmail-103.patch

Sometimes (rarely) the answer to a DNS query is larger than 512 bytes.
This is an important limit: it's the biggest response (unpatched) qmail
can understand. This threshold was originally based upon the UDP DNS
protocol definition (RFC 1025, section 4.2.1). It is still relatively
unusual for DNS responses to be bigger than that limit, even though they
are allowed to be. However, it does happen from time to time,
particularly on domains with a lot of alternatives (if each MX record is
approximately 50 bytes, you need 11 of them to cross the limit). This
patch allows DNS packets to be as big as the maximum EDNS response size:
64kb (though RFC 6891 strongly recommends capping responses at 4kb).

(Description from http://www.memoryhole.net/qmail/#oversize-dns)