Add qmail-rcptcheck(8) from rejectutils.
qmail-rcptcheck runs an administrator-defined sequence of programs to
check SMTP envelope senders and recipients. Checks must adhere to the
RCPTCHECK interface. If any check rejects, the message is rejected.

qmail-rcptcheck is most commonly invoked from qmail-smtpd via RCPTCHECK,
but also runs under qmail-spp (not documented here), so if and when we
switch to that, recipient-checking programs won't need to be changed.