qmail-qstat: simplify the code
First, do not chdir to the qmail directory, but instead pass the absolute path
to the find commands.

Then tell find to only output files, and count them, instead of counting
everything and then substracting the directories. If there is anything beyond
a plain file somewhere in the queue this would have been counted before and is
no longer counted now, but the presence of anything else can be considered a
bug in itself.

When someone is using the big-todo patch this would collide, but for good: the
new version can work with both the original and the big-todo queue style
without modifications.

While at it quote the arguments to echo just to be sure.

The additional "-type f" argument can already be found in the AT&T System V
user manual from 1986, so it's safe to assume it's presence.