Simply triggering builds on push is sufficient.
We expect merging a PR to run autobuilds on the target branch (e.g.,
master). The recent autobuild improvements caused a regression here:
merging a PR has lately been running autobuilds on the *source* branch.
So GitHub has no information to offer about build status on master,
which also confounds the status badge recently added to README.md.

As we suspected, I had overengineered the trigger constraints. Merging a
PR causes the target branch to receive a push, we don't need to treat
any branches (or PRs!) specially in autobuilds, and the branch
protection rules already in place for master -- together with running
builds on every branch push -- are sufficient for PR merges to be as
safe as they've always been.