substdio: make substdio_put*s() static inlines
These functions are usually called with a character literal as second argument.
When these functions are static inlines this allows the compiler to entirely
avoid the runtime call to strlen() and pass a constant there. These functions
are not turned into defines as that may evaluate the "buf" argument twice which
would suddenly emit incorrect code if someone passes things like "buf++".