qmail-qmqpc: set the correct fd in the substdio structs
The "to" and "from" substdios are initialized with fd -1 because the socket is
dynamically allocated at runtime. That is why the passed fd in the saferead()
and safewrite() functions was ignored and the socket descriptor was always used.
Update the fd in the substdios once the socket is allocated and then use the
provided fd.